+++ Adj/Program/Primitive/Generation.hs
@@ -4,7 +4,7 @@
 module Adj.Program.Primitive.Generation where
 
 import Adj.Auxiliary (Casted, Casting ((=-), (-=)), type (=!?=), FG (FG), FFGH (FFGH), type (=!!??=), Structural (Structural))
-import Adj.Algebra.Category (Semigroupoid ((.)), Category ((.:), (...:), (....:), identity), Functor (map), Covariant, Bindable, Traversable, Semimonoidal, Component (component), Identity (Identity), Day (Day), type (-->), type (-/->), Straight (Straight), Opposite, Kleisli (Kleisli), (->-), (->>-), (->>--), (->--), (-->--), (-/>/-), (-/>>/-), (=-=))
+import Adj.Algebra.Category (Semigroupoid ((.)), Category ((.:), (...:), identity), Functor (map), Covariant, Bindable, Traversable, Semimonoidal, Component (component), Identity (Identity), Day (Day), type (-->), type (-/->), Straight (Straight), Opposite, Kleisli (Kleisli), (->-), (->>-), (->>--), (->--), (-->--), (-/>/-), (-/>>/-), (=-=))
 import Adj.Algebra.Set (Setoid, (:*:) ((:*:)), (:+:) (This, That))
 
 newtype Generation f g o = Generation
@@ -18,6 +18,9 @@ deriving via (Structural ((=!!??=) f Identity (g =!?= Generation f g) o))
 instance Casting (->) (Generation f g) where
 	(=-) (Generation m) = m
 	(-=) m = Generation m
+
+pattern Generate :: f (Identity o) (FG g (Generation f g) o) -> Generation f g o
+pattern Generate xs <- Generation (FFGH xs) where Generate xs = Generation (FFGH xs)
 
 instance
 	( forall o . Functor (-->) (-->) (Straight f o)
@@ -42,19 +45,18 @@ instance
 type Construction = Generation (:*:)
 
 pattern Construct :: o -> f (Construction f o) -> Construction f o
-pattern Construct x xs <- Generation (FFGH (Identity x :*: FG xs))
-	where Construct x xs = Generation . FFGH ...: Identity x :*: FG xs
+pattern Construct x xs <- Generate (Identity x :*: FG xs)
+	where Construct x xs = Generate ...: Identity x :*: FG xs
 
 type Instruction = Generation (:+:)
 
 pattern Instruct :: f (Instruction f o) -> Instruction f o
-pattern Instruct xs <- Generation (FFGH (That (FG xs)))
-	where Instruct xs = Generation . FFGH . That .: FG xs
+pattern Instruct xs <- Generate (That (FG xs))
+	where Instruct xs = Generate . That .: FG xs
 
 pattern Load :: o -> Instruction f o
-pattern Load x <- Generation (FFGH (This (Identity x)))
-	where Load x = Generation . FFGH . This .: Identity x
+pattern Load x <- Generate (This (Identity x))
+	where Load x = Generate . This .: Identity x
 
 instance Covariant Straight Functor f (->) (->) => Component (-->) f (Instruction f) where
-	component = Straight .: \x -> Generation . FFGH . That . FG
-		....: Generation . FFGH . This . Identity ->- x
+	component = Straight .: \x -> Instruct ...: Load ->- x